% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse.R
\name{read_synapse}
\alias{read_synapse}
\title{Read flyem format synapse.json file}
\usage{
read_synapse(x)
}
\arguments{
\item{x}{The path to a flyem synapse.json file (or a URL)}
}
\value{
A list containing two \code{dataframe}s: \itemize{

  \item{\code{tbars}} containing information about presynaptic T Bars

  \item{\code{partners}} detailing the postsynaptic partners for each T bar. }
}
\description{
Read flyem format synapse.json file
}
\examples{
f=system.file('testdata/testsynapse.json', package = 'rflyem')
s <- read_synapse(f)
summary(s$partners)
plot(s$partners[,c("x","y")], asp=1)
}
